refactor: split ai layout service tests
